ModelReflexRobot (Incomplete, Runs)
This has a model based reflex robot, which contains an internal state that it uses to navigate the room. As moves around, the state updates itself. It is supposed to prioritize moving towards dirt over unexplored tiles, but it does the opposite. I think this has something to do with the findClosest method. Maybe the method is not searching the robot's internal state correctly.